% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice.R
\name{count_na}
\alias{count_na}
\title{Count number of NA values in a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The number of NA values in the vector
}
\description{
Count number of NA values in a vector
}
